noam - refactored to cleaner code solutoin
