+++ src/main/java/edu/greenblitz/tobyDetermined/Robot.java
@@ -9,8 +9,8 @@ import edu.greenblitz.tobyDetermined.subsystems.Dashboard;
 import edu.greenblitz.tobyDetermined.subsystems.LED;
 import edu.greenblitz.tobyDetermined.subsystems.Limelight.MultiLimelight;
 import edu.greenblitz.tobyDetermined.subsystems.RotatingBelly.RotatingBelly;
-import edu.greenblitz.tobyDetermined.subsystems.intake.IntakeExtender;
-import edu.greenblitz.tobyDetermined.subsystems.intake.IntakeRoller;
+import edu.greenblitz.tobyDetermined.subsystems.swerve.intake.IntakeExtender;
+import edu.greenblitz.tobyDetermined.subsystems.swerve.intake.IntakeRoller;
 import edu.greenblitz.tobyDetermined.subsystems.swerve.SwerveChassis;
 import edu.greenblitz.tobyDetermined.subsystems.telescopicArm.ArmSimulation;
 import edu.greenblitz.tobyDetermined.subsystems.telescopicArm.Claw;
