+++ src/main/java/edu/greenblitz/tobyDetermined/commands/Auto/balance/AdvancedBalanceOnRamp.java
@@ -1,7 +1,7 @@
 package edu.greenblitz.tobyDetermined.commands.Auto.balance;
 
 import edu.greenblitz.tobyDetermined.commands.swerve.SwerveCommand;
-import edu.greenblitz.tobyDetermined.subsystems.DriveTrain.IO.GyroIOs.IGyro;
+import edu.greenblitz.tobyDetermined.subsystems.GyroIOs.IGyro;
 import edu.greenblitz.utils.PitchRollAdder;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
