+++ src/main/java/edu/greenblitz/tobyDetermined/commands/Auto/balance/LockWheels.java
@@ -2,7 +2,7 @@ package edu.greenblitz.tobyDetermined.commands.Auto.balance;
 
 import edu.greenblitz.tobyDetermined.commands.swerve.RotateAllWheelsToAngle;
 import edu.greenblitz.tobyDetermined.commands.swerve.SwerveCommand;
-import edu.greenblitz.tobyDetermined.subsystems.DriveTrain.IO.GyroIOs.IGyro;
+import edu.greenblitz.tobyDetermined.subsystems.GyroIOs.IGyro;
 
 // import edu.wpi.first.math.geometry.Rotation2d;
 public class LockWheels extends SwerveCommand {
