+++ src/main/java/edu/greenblitz/tobyDetermined/commands/intake/extender/ExtenderCommand.java
@@ -1,6 +1,6 @@
 package edu.greenblitz.tobyDetermined.commands.intake.extender;
 
-import edu.greenblitz.tobyDetermined.subsystems.intake.IntakeExtender;
+import edu.greenblitz.tobyDetermined.subsystems.swerve.intake.IntakeExtender;
 import edu.greenblitz.utils.GBCommand;
 
 public abstract class ExtenderCommand extends GBCommand {
