+++ src/main/java/edu/greenblitz/tobyDetermined/commands/intake/roller/RollerCommand.java
@@ -1,6 +1,6 @@
 package edu.greenblitz.tobyDetermined.commands.intake.roller;
 
-import edu.greenblitz.tobyDetermined.subsystems.intake.IntakeRoller;
+import edu.greenblitz.tobyDetermined.subsystems.swerve.intake.IntakeRoller;
 import edu.greenblitz.utils.GBCommand;
 
 public abstract class RollerCommand extends GBCommand {
