+++ src/main/java/edu/greenblitz/tobyDetermined/subsystems/GyroIOs/IGyro.java
@@ -1,7 +1,6 @@
-package edu.greenblitz.tobyDetermined.subsystems.DriveTrain.IO.GyroIOs;
+package edu.greenblitz.tobyDetermined.subsystems.GyroIOs;
 
-import edu.greenblitz.tobyDetermined.subsystems.DriveTrain.IO.inputs.GyroInputsAutoLogged;
-import org.littletonrobotics.junction.AutoLog;
+import edu.greenblitz.tobyDetermined.subsystems.swerve.inputs.GyroInputsAutoLogged;
 
 public interface IGyro {
