+++ src/main/java/edu/greenblitz/tobyDetermined/subsystems/swerve/ModuleIOs/ISwerveModule.java
@@ -1,7 +1,6 @@
-package edu.greenblitz.tobyDetermined.subsystems.DriveTrain.IO.ModuleIOs;
+package edu.greenblitz.tobyDetermined.subsystems.swerve.ModuleIOs;
 
-import edu.greenblitz.tobyDetermined.subsystems.DriveTrain.IO.inputs.SwerveModuleInputsAutoLogged;
-import org.littletonrobotics.junction.AutoLog;
+import edu.greenblitz.tobyDetermined.subsystems.swerve.inputs.SwerveModuleInputsAutoLogged;
 
 public interface ISwerveModule {
